ignore first delete

+++ serveImage.js
@@ -11,12 +11,15 @@ const output = './images/image.jpg';
 
 function generateImage(path, int){
     const pre = './tmp/image'+(int-1);
-    try {
-        fs.unlinkSync(pre+"-resized-height.jpg")
-        fs.unlinkSync(pre+"-resized-width.jpg")
-        fs.unlinkSync(pre+"-resized-width-cropped.jpg")
-    } catch(err) {
-        console.error(err)
+    console.log(pre);
+    if(!pre===0){
+        try {
+            fs.unlinkSync(pre+"-resized-height.jpg")
+            fs.unlinkSync(pre+"-resized-width.jpg")
+            fs.unlinkSync(pre+"-resized-width-cropped.jpg")
+        } catch(err) {
+            console.error(err)
+        }
     }
     const tmp = './tmp/image'+int;
     meta.getMetadata(path).then(x => {
